check if pid file exists before attempting to delete, add logging

+++ fish/functions/fq_hm.fish
@@ -1,8 +1,12 @@
 function fq_hm --description "Forcefully quits rails"
-    # TODO: Check if file exists first before attempting to remove
-    # TODO: Add logging that explains what is happening at each step
     # TODO: De-hardcode path to hm
-    rm /Users/drew.daniels/projects/sites/healthmatters/tmp/pids/server.pid
+    # TODO: Put this into a variable
+    if test -e /Users/drew.daniels/projects/sites/healthmatters/tmp/pids/server.pid
+        echo "Removing Rails Server PID File"
+        rm /Users/drew.daniels/projects/sites/healthmatters/tmp/pids/server.pid
+    end
+    echo "Killing all rails processes"
     ps -ef | grep rails | grep -v grep | awk '{print $2}' | xargs kill -9
+    echo "Killing all sidekiq processes"
     ps -ef | grep sidekiq | grep -v grep | awk '{print $2}' | xargs kill -9
 end
